+++ HISTORY.md
@@ -9,6 +9,7 @@ v1.5.0 ()
 * Using OS/Version/CPU_ARCH for platform information
 * Fixed serial port closes issue
 * Changed default device meta manufactor to FreeIOE
+* Support upgrade skynet code without freeioe